Split visibility from FormItem class and create subclass of FormItem with visibility
