+++ src/main/java/com/karankumar/bookproject/ui/book/components/form/item/PagesRead.java
@@ -4,7 +4,7 @@ import com.karankumar.bookproject.backend.entity.Book;
 import com.vaadin.flow.component.textfield.IntegerField;
 import com.vaadin.flow.data.binder.Binder;
 
-public class PagesRead extends FormItem<IntegerField> {
+public class PagesRead extends VisibleFormItem<IntegerField> {
 
     public PagesRead() {
         super(new IntegerField());
