+++ src/main/java/com/karankumar/bookproject/ui/book/components/form/item/VisibleFormItem.java
@@ -1,0 +1,30 @@
+package com.karankumar.bookproject.ui.book.components.form.item;
+
+import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.formlayout.FormLayout;
+
+public abstract class VisibleFormItem<T extends Component> extends FormItem<T>{
+    private FormLayout.FormItem item;
+
+    public VisibleFormItem(T field) {
+        super(field);
+    }
+
+    @Override
+    public void add(FormLayout layout) {
+        item = layout.addFormItem(super.getField(), getLabel());
+    }
+
+    public void show() {
+        item.setVisible(true);
+    }
+
+    public void hide() {
+        item.setVisible(false);
+    }
+
+    public boolean isVisible() {
+        return item.isVisible();
+    }
+
+}
